requirement.txt
asgiref==3.5.0
bleach==4.1.0
Django==4.0.2
django-filter==21.1
django-summernote==0.8.20.0
gunicorn==20.1.0
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
webencodings==0.5.1
whitenoise==6.0.0
